+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'Qiitq'
+rootProject.name = 'Qiita'
