+++ src/main/kotlin/com/final/project/Qiita/controller/UserController.kt
@@ -1,15 +1,15 @@
-package com.final.project.Qiitq.controller
+package com.final.project.Qiita.controller
 
 import org.springframework.stereotype.Controller
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 
 @Controller
-@RequestMapping("/login")
-class LoginController {
+@RequestMapping("/user")
+class UserController {
 
     @GetMapping("/", "")
-    fun login(): String {
+    fun show(): String {
         return "login"
     }
 }
