+++ src/main/kotlin/com/final/project/Qiita/domain/User.kt
@@ -1,4 +1,4 @@
-package com.final.project.Qiitq.domain
+package com.final.project.Qiita.domain
 
 class User(var name: String, var email: String, var password: String, var description: String) {
     var id: Int = 0
