+++ src/main/resources/com/final/project/Qiita/mapper/UserMapper.xml
@@ -2,16 +2,16 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.final.project.Qiitq.mapper.UserMapper">
+<mapper namespace="com.final.project.Qiita.mapper.UserMapper">
     <insert id="insert" useGeneratedKeys="true" keyProperty="id">
         INSERT INTO user (name, email, password, description) VALUES (#{name}, #{email}, #{password}, #{description})
     </insert>
 
-    <select id="select" resultType="com.final.project.Qiitq.domain.User">
+    <select id="select" resultType="com.final.project.Qiita.domain.User">
         SELECT * FROM user WHERE id = #{id}
     </select>
 
-    <select id="all" resultType="com.final.project.Qiitq.domain.User">
+    <select id="all" resultType="com.final.project.Qiita.domain.User">
         SELECT * FROM user;
     </select>
